+++ src/splat/util/symbols.py
@@ -235,11 +235,11 @@ def handle_sym_addrs(
                             if attr_name == "dont_allow_addend":
                                 sym.dont_allow_addend = tf_val
                                 continue
-                            if attr_name == "allow_reference":
-                                sym.allow_reference = tf_val
+                            if attr_name == "can_reference":
+                                sym.can_reference = tf_val
                                 continue
-                            if attr_name == "allow_be_referenced":
-                                sym.allow_be_referenced = tf_val
+                            if attr_name == "can_be_referenced":
+                                sym.can_be_referenced = tf_val
                                 continue
                             if attr_name == "allow_duplicated":
                                 sym.allow_duplicated = True
@@ -484,10 +484,10 @@ def add_symbol_to_spim_segment(
         context_sym.allowedToReferenceAddends = True
     if sym.dont_allow_addend:
         context_sym.notAllowedToReferenceAddends = True
-    if sym.allow_reference is not None:
-        context_sym.allowedToReferenceSymbols = sym.allow_reference
-    if sym.allow_be_referenced is not None:
-        context_sym.allowedToBeReferenced = sym.allow_be_referenced
+    if sym.can_reference is not None:
+        context_sym.allowedToReferenceSymbols = sym.can_reference
+    if sym.can_be_referenced is not None:
+        context_sym.allowedToBeReferenced = sym.can_be_referenced
     context_sym.setNameGetCallbackIfUnset(lambda _: sym.name)
     if sym.given_name_end:
         context_sym.nameEnd = sym.given_name_end
@@ -630,8 +630,8 @@ class Symbol:
     allow_addend: bool = False
     dont_allow_addend: bool = False
 
-    allow_reference: Optional[bool] = None
-    allow_be_referenced: Optional[bool] = None
+    can_reference: Optional[bool] = None
+    can_be_referenced: Optional[bool] = None
 
     linker_section: Optional[str] = None